add Feat FoodSearchResult.

+++ app/src/main/java/com/chs/your_body_profile/domain/model/FoodDetailInfo.kt
@@ -3,14 +3,14 @@ package com.chs.your_body_profile.domain.model
 
 data class FoodDetailInfo(
     val name: String,
-    val servingWeight: Float,
-    val calorie: Int,
-    val carbohydrate: Float,
-    val fat: Float,
-    val protein: Float,
-    val sodium: Float,
-    val sugar: Float,
-    val cholesterol: Float,
-    val saturatedFat: Float,
-    val transFat: Float
+    val servingWeight: Double,
+    val calorie: Double,
+    val carbohydrate: Double,
+    val fat: Double,
+    val protein: Double,
+    val sodium: Double,
+    val sugar: Double,
+    val cholesterol: Double,
+    val saturatedFat: Double,
+    val transFat: Double
 )
